# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user/refresh-token": {
      "post": {
        "operationId": "UserController_refreshToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Access new refresh token"
          }
        },
        "summary": "Refreshing new token",
        "tags": [
          "User"
        ]
      }
    },
    "/user/refresh-token-google": {
      "post": {
        "operationId": "UserController_refreshTokenGoogle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Access new refresh token"
          }
        },
        "summary": "Refreshing new token",
        "tags": [
          "User"
        ]
      }
    },
    "/user/register": {
      "post": {
        "operationId": "UserController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully"
          }
        },
        "summary": "Register a new user",
        "tags": [
          "User"
        ]
      }
    },
...
    "/user/all": {
      "get": {
        "operationId": "UserController_getAllUsers",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "end_date",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "start_date",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "role_code",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "department_code",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "staff_id",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "email",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "full_name",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "List of users"
          }
        },
        "summary": "Get all users with filters",
        "tags": [
          "User"
        ]
      }
    },
...
    "/user/me/change-password": {
      "post": {
        "operationId": "UserController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully"
          }
        },
        "summary": "Change current user password",
        "tags": [
          "User"
        ]
      }
    },
    "/user/{id}": {
      "get": {
        "operationId": "UserController_getUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details"
          }
        },
        "summary": "Get a user by ID",
        "tags": [
          "User"
        ]
      },
      "patch": {
        "operationId": "UserController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          }
        },
        "summary": "Update a user",
        "tags": [
          "User"
        ]
      },
      "delete": {
        "operationId": "UserController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully"
          }
        },
        "summary": "Delete a user",
        "tags": [
          "User"
        ]
      }
    },
...
  },
  "info": {
    "title": "Travel Booking API Documentation",
    "description": "API documentation for Travel Booking",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt": {
        "scheme": "Bearer",
        "bearerFormat": "Bearer",
        "description": "[just text field] Please enter token in following format: Bearer <JWT>",
        "name": "Authorization",
        "type": "http",
        "in": "Header"
      }
    },
    "schemas": {
...
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "johndoe@gmail.com"
          },
          "staff_id": {
            "type": "string",
            "example": "user123"
          },
          "full_name": {
            "type": "string",
            "example": "John Doe"
          },
          "password": {
            "type": "string",
            "example": "securepassword"
          },
          "confirm_password": {
            "type": "string",
            "example": "securepassword"
          },
          "status": {
            "type": "string",
            "example": "Active",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "PENDING",
              "APPROVED",
              "REJECTED",
              "DELETE"
            ],
            "default": "ACTIVE"
          },
          "user_type": {
            "type": "string",
            "example": "Maker",
            "enum": [
              "MAKER",
              "CHECKER"
            ],
            "default": "MAKER"
          },
          "department_code": {
            "type": "string",
            "example": "dept01"
          },
          "role_code": {
            "type": "string",
            "example": "role01"
          }
        },
        "required": [
          "email",
          "staff_id",
          "full_name",
          "password",
          "confirm_password",
          "status",
          "user_type",
          "department_code",
          "role_code"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "current_password": {
            "type": "string",
            "example": "currentpassword123"
          },
          "new_password": {
            "type": "string",
            "example": "newpassword123"
          },
          "confirm_new_password": {
            "type": "string",
            "example": "newpassword123"
          }
        },
        "required": [
          "current_password",
          "new_password",
          "confirm_new_password"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "johndoe@gmail.com"
          },
          "staff_id": {
            "type": "string",
            "example": "user123"
          },
          "full_name": {
            "type": "string",
            "example": "John Doe"
          },
          "password": {
            "type": "string",
            "example": "securepassword"
          },
          "confirm_password": {
            "type": "string",
            "example": "securepassword"
          },
          "status": {
            "type": "string",
            "example": "Active",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "PENDING",
              "APPROVED",
              "REJECTED",
              "DELETE"
            ],
            "default": "ACTIVE"
          },
          "user_type": {
            "type": "string",
            "example": "Maker",
            "enum": [
              "MAKER",
              "CHECKER"
            ],
            "default": "MAKER"
          },
          "department_code": {
            "type": "string",
            "example": "dept01"
          },
          "role_code": {
            "type": "string",
            "example": "role01"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "jwt": []
    }
  ]
}